backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-sinmungo-backend"
version = "0.1.0"
description = "AI 신문고 백엔드"
authors = [{name = "최지안"}]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "instructor>=0.4.0",
    "langchain>=0.1.0",
    "langgraph>=0.0.20",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.1.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "openai-whisper>=20231117",
    "anthropic>=0.7.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "pillow>=12.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
